ISSUE-16 - Refactor further into separate modules, nesting several modules which are not required by the rest of the program.
* Split out several more modules from top level format module
* Split scan module into several sub modules
* Add new args module
