+++ cargo-geiger/src/format/table.rs
@@ -1,9 +1,8 @@
-use crate::find::{unsafe_stats, GeigerContext};
+use crate::format::emoji_symbols::EmojiSymbols;
 use crate::format::print::{colorize, PrintConfig};
-use crate::format::tree::TextTreeLine;
-use crate::format::{
-    get_kind_group_name, CrateDetectionStatus, EmojiSymbols, SymbolKind,
-};
+use crate::format::{get_kind_group_name, CrateDetectionStatus, SymbolKind};
+use crate::scan::{unsafe_stats, GeigerContext};
+use crate::tree::TextTreeLine;
 
 use cargo::core::package::PackageSet;
 use geiger::{Count, CounterBlock};
@@ -42,36 +41,45 @@ pub fn create_table_from_text_tree_lines(
                     // TODO: Avoid panic, return Result.
                     panic!("Expected to find package by id: {}", id);
                 });
-                let pack_metrics = match geiger_context.pack_id_to_metrics.get(&id) {
-                    Some(m) => m,
-                    None => {
-                        warning_count += package_is_new as u64;
-                        eprintln!("WARNING: No metrics found for package: {}", id);
-                        continue;
-                    }
-                };
+                let pack_metrics =
+                    match geiger_context.pack_id_to_metrics.get(&id) {
+                        Some(m) => m,
+                        None => {
+                            warning_count += package_is_new as u64;
+                            eprintln!(
+                                "WARNING: No metrics found for package: {}",
+                                id
+                            );
+                            continue;
+                        }
+                    };
                 let unsafety = unsafe_stats(pack_metrics, rs_files_used);
                 if package_is_new {
-                    total_package_counts.total_counter_block += unsafety.used.clone();
-                    total_package_counts.total_unused_counter_block += unsafety.unused.clone();
+                    total_package_counts.total_counter_block +=
+                        unsafety.used.clone();
+                    total_package_counts.total_unused_counter_block +=
+                        unsafety.unused.clone();
                 }
                 let unsafe_found = unsafety.used.has_unsafe();
                 let crate_forbids_unsafe = unsafety.forbids_unsafe;
                 let total_inc = package_is_new as i32;
-                let crate_detection_status = match (unsafe_found, crate_forbids_unsafe) {
-                    (false, true) => {
-                        total_package_counts.none_detected_forbids_unsafe += total_inc;
-                        CrateDetectionStatus::NoneDetectedForbidsUnsafe
-                    }
-                    (false, false) => {
-                        total_package_counts.none_detected_allows_unsafe += total_inc;
-                        CrateDetectionStatus::NoneDetectedAllowsUnsafe
-                    }
-                    (true, _) => {
-                        total_package_counts.unsafe_detected += total_inc;
-                        CrateDetectionStatus::UnsafeDetected
-                    }
-                };
+                let crate_detection_status =
+                    match (unsafe_found, crate_forbids_unsafe) {
+                        (false, true) => {
+                            total_package_counts
+                                .none_detected_forbids_unsafe += total_inc;
+                            CrateDetectionStatus::NoneDetectedForbidsUnsafe
+                        }
+                        (false, false) => {
+                            total_package_counts.none_detected_allows_unsafe +=
+                                total_inc;
+                            CrateDetectionStatus::NoneDetectedAllowsUnsafe
+                        }
+                        (true, _) => {
+                            total_package_counts.unsafe_detected += total_inc;
+                            CrateDetectionStatus::UnsafeDetected
+                        }
+                    };
 
                 let icon = match crate_detection_status {
                     CrateDetectionStatus::NoneDetectedForbidsUnsafe => {
@@ -210,10 +218,7 @@ fn table_footer(
     colorize(output, &status)
 }
 
-fn table_row(
-    used: &CounterBlock,
-    not_used: &CounterBlock,
-) -> String {
+fn table_row(used: &CounterBlock, not_used: &CounterBlock) -> String {
     let fmt = |used: &Count, not_used: &Count| {
         format!("{}/{}", used.unsafe_, used.unsafe_ + not_used.unsafe_)
     };
@@ -228,7 +233,8 @@ fn table_row(
 }
 
 fn table_row_empty() -> String {
-    let headers_but_last = &UNSAFE_COUNTERS_HEADER[..UNSAFE_COUNTERS_HEADER.len() - 1];
+    let headers_but_last =
+        &UNSAFE_COUNTERS_HEADER[..UNSAFE_COUNTERS_HEADER.len() - 1];
     let n = headers_but_last
         .iter()
         .map(|s| s.len())
