+++ cargo-geiger/src/scan/find.rs
@@ -1,0 +1,150 @@
+use crate::format::print::PrintConfig;
+use crate::rs_file::{
+    into_rs_code_file, is_file_with_ext, PackageMetrics, RsFile,
+    RsFileMetricsWrapper,
+};
+
+use super::{GeigerContext, ScanMode};
+
+use cargo::core::package::PackageSet;
+use cargo::core::{Package, PackageId};
+use cargo::util::CargoResult;
+use cargo::{CliError, Config};
+use geiger::{find_unsafe_in_file, IncludeTests};
+use std::collections::HashMap;
+use std::path::Path;
+use std::path::PathBuf;
+use walkdir::WalkDir;
+
+pub fn find_unsafe(
+    mode: ScanMode,
+    config: &Config,
+    packages: &PackageSet,
+    print_config: &PrintConfig,
+) -> Result<GeigerContext, CliError> {
+    let mut progress = cargo::util::Progress::new("Scanning", config);
+    let geiger_context = find_unsafe_in_packages(
+        packages,
+        print_config.allow_partial_results,
+        print_config.include_tests,
+        mode,
+        |i, count| -> CargoResult<()> { progress.tick(i, count) },
+    );
+    progress.clear();
+    config.shell().status("Scanning", "done")?;
+    Ok(geiger_context)
+}
+
+fn find_unsafe_in_packages<F>(
+    packs: &PackageSet,
+    allow_partial_results: bool,
+    include_tests: IncludeTests,
+    mode: ScanMode,
+    mut progress_step: F,
+) -> GeigerContext
+where
+    F: FnMut(usize, usize) -> CargoResult<()>,
+{
+    let mut pack_id_to_metrics = HashMap::new();
+    let packs = packs.get_many(packs.package_ids()).unwrap();
+    let pack_code_files: Vec<_> = find_rs_files_in_packages(&packs).collect();
+    let pack_code_file_count = pack_code_files.len();
+    for (i, (pack_id, rs_code_file)) in pack_code_files.into_iter().enumerate()
+    {
+        let (is_entry_point, p) = match rs_code_file {
+            RsFile::LibRoot(pb) => (true, pb),
+            RsFile::BinRoot(pb) => (true, pb),
+            RsFile::CustomBuildRoot(pb) => (true, pb),
+            RsFile::Other(pb) => (false, pb),
+        };
+        if let (false, ScanMode::EntryPointsOnly) = (is_entry_point, &mode) {
+            continue;
+        }
+        match find_unsafe_in_file(&p, include_tests) {
+            Err(e) => {
+                if allow_partial_results {
+                    eprintln!(
+                        "Failed to parse file: {}, {:?} ",
+                        &p.display(),
+                        e
+                    );
+                } else {
+                    panic!("Failed to parse file: {}, {:?} ", &p.display(), e);
+                }
+            }
+            Ok(file_metrics) => {
+                let package_metrics = pack_id_to_metrics
+                    .entry(pack_id)
+                    .or_insert_with(PackageMetrics::default);
+                let wrapper = package_metrics
+                    .rs_path_to_metrics
+                    .entry(p)
+                    .or_insert_with(RsFileMetricsWrapper::default);
+                wrapper.metrics = file_metrics;
+                wrapper.is_crate_entry_point = is_entry_point;
+            }
+        }
+        let _ = progress_step(i, pack_code_file_count);
+    }
+    GeigerContext { pack_id_to_metrics }
+}
+
+fn find_rs_files_in_dir(dir: &Path) -> impl Iterator<Item = PathBuf> {
+    let walker = WalkDir::new(dir).into_iter();
+    walker.filter_map(|entry| {
+        let entry = entry.expect("walkdir error."); // TODO: Return result.
+        if !is_file_with_ext(&entry, "rs") {
+            return None;
+        }
+        Some(
+            entry
+                .path()
+                .canonicalize()
+                .expect("Error converting to canonical path"),
+        ) // TODO: Return result.
+    })
+}
+
+fn find_rs_files_in_package(pack: &Package) -> Vec<RsFile> {
+    // Find all build target entry point source files.
+    let mut canon_targets = HashMap::new();
+    for t in pack.targets() {
+        let path = t.src_path().path();
+        let path = match path {
+            None => continue,
+            Some(p) => p,
+        };
+        if !path.exists() {
+            // A package published to crates.io is not required to include
+            // everything. We have to skip this build target.
+            continue;
+        }
+        let canon = path
+            .canonicalize() // will Err on non-existing paths.
+            .expect("canonicalize for build target path failed."); // FIXME
+        let targets = canon_targets.entry(canon).or_insert_with(Vec::new);
+        targets.push(t);
+    }
+    let mut out = Vec::new();
+    for p in find_rs_files_in_dir(pack.root()) {
+        if !canon_targets.contains_key(&p) {
+            out.push(RsFile::Other(p));
+        }
+    }
+    for (k, v) in canon_targets.into_iter() {
+        for target in v {
+            out.push(into_rs_code_file(target.kind(), k.clone()));
+        }
+    }
+    out
+}
+
+fn find_rs_files_in_packages<'a>(
+    packs: &'a [&Package],
+) -> impl Iterator<Item = (PackageId, RsFile)> + 'a {
+    packs.iter().flat_map(|pack| {
+        find_rs_files_in_package(pack)
+            .into_iter()
+            .map(move |path| (pack.package_id(), path))
+    })
+}
